requirements.txt
requests
psycopg2-binary
python-dateutil
selenium
pillow
webdriver-manager
python-dotenv